python/requirements.txt
tensorflow==2.10.0
numpy==1.23.5
pandas==1.5.2
scikit-learn==1.2.0
matplotlib==3.6.2
seaborn==0.12.2
pillow==9.3.0
opencv-python==4.7.0.68
flask==2.2.2
flask-cors==3.0.10
tqdm==4.64.1

